Add swh.webhooks.Webhooks class and event types management methods
Add the swh.webhooks.Webhooks class wrapping requests to the Svix
REST API. Svix is an opensource framework designed to ease webhooks
management.

Define methods used to manage webhook event types, the following
operations are supported:

- creating or updating an event type

- retrieving a specific event type

- listing all registered event types

- deleting an event type

A webhook event type is defined by a name, a description and a JSON
schema detailing the content of the payload sent when an event is
triggered.

Related to #2