swh.webhooks.Webhooks: Add method to send event to endpoints
Add a method to send an event of a given type with a JSON payload
to registered endpoints.

The event can also be sent to a specific channel, only endpoints
listening to this channel will receive it.

Before sending an event, the payload content is checked against
the JSON schema of the event and an exception is raised if the
validation fails.

Related to #4