cli: Add webhooks command group and initialize Webhooks instance
Add the "webhooks" group to the main swh CLI and add options enabling to
initialize the Webhooks instance by providing authentication token and
URL for the Svix server.

Those values can either be provided through a configuration file or by
explicitly passing them as parameters of the "swh webhooks" command group.

Related to #6.